refactor(db): extrai schema-column reconciliation de db/core.ts
Continua a decomposição do god-file db/core.ts (empilhado sobre #4947). Move
os reconciliadores idempotentes de coluna (ensureProviderConnectionsColumns,
ensureUsageHistoryColumns, ensureCallLogsColumns) + os helpers de introspecção
em que se apoiam (hasColumn, hasTable, quoteIdentifier, getTableColumns) para
db/schemaColumns.ts.

- Cada um recebe o db handle explicitamente — sem estado de módulo → leaf
  co-localizado que core.ts chama no bootstrap de getDbInstance().
- core.ts reimporta os 6 usados internamente (ensure* no bootstrap; hasTable/
  quoteIdentifier/getTableColumns em captureCriticalDbState/restore). hasColumn
  não tinha uso interno nem externo (cada consumidor externo tem def local) —
  movido ao leaf e exportado lá; host não o importa.

Behavior-preserving (byte-idêntico). core.ts 1715 → 1519.

Validação: typecheck:core, eslint 0/0, check:cycles, check:file-size +
teste de caracterização novo (3/3, DB :memory: real exercitando idempotência
do ensure* e introspecção).
